+++ sales-Backend/sales-Backend/src/main/java/com/example/salesBackend/Entity/PG_RECEIPTS.java
@@ -4,6 +4,8 @@ package com.example.salesBackend.Entity;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.GenericGenerator;
+
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -15,11 +17,10 @@ import java.util.Date;
 @Table(name="PG_RECEIPTS")
 public class PG_RECEIPTS {
     @Id
-//    @GeneratedValue(strategy=GenerationType.UUID)
     @Column(name = "POLICY_NO")
     private String POLICY_NO;
 
     private String RECEIPT_NO;
     private Date RECEIPT_DATE;
-    private Long AMOUNT;
+    private BigDecimal AMOUNT;
 }
